+++ README.md
@@ -102,7 +102,6 @@ Execute command in the Project Root folder
     @pmm-ami                legacy/deprecated
     @not-ui-pipeline        legacy/deprecated
     @not-pr-pipeline        legacy/deprecated
-    @cli                    cli related tests
 
 
 ## Contributing
